agressive batching
We are doing batching in these places.

1. Instead of flush on every write (of a coutinuation), we buffer them
and Flush() when a level of continuation is about to finish.

2. Before a continuation is sent to the schedulers, we buffer them in
a per-core local buffer. We push to the schedulers either when the
local buffer is full or an entire level is finished. (See
PromiseRoundRobinImpl)

3. Before sending continuations to the global TCP stream, we buffer
them into a local buffer too. The last two tricks can greatly reduce
lock contention.