Reduce memory footprint to be able to run on thin nodes
This commit brings down the memory usage of felis to around ~16 GB
when running the TPC-C workload with 4 warehouses on 4 cores.

A tracker for memory usage has been added, which groups allocations
by what they are for. Allocations made through mem::Pool have extra
instrumentation, which collects the current usage and peak usage.

Statistics are currently outputted at the end of each epoch.