Fix double join issue on exit
caused by calling go::WaitThreadPool() once in main.cc, and
separately in CoroutineModule::~CoroutineModule() upon destruction
of the static CoroutineModule object.

Waiting for the coroutines has to happen in main (and not in
the destructor) since the static CoroutineModule object's destructor
has no guarantee to be called before that of the static std::vector
containing the 'std::thread's.